// src/app/@api/model/supplierEmployeeDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SupplierDto } from './supplierDto';
import { UserRequestDto } from './userRequestDto';


export interface SupplierEmployeeDto { 
    supplier: SupplierDto;
    user: UserRequestDto;
}