// src/app/@api/model/updateSupplierProfileDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UpdateSupplierSupplierDto } from './updateSupplierSupplierDto';
import { UpdateSupplierUserProfileDto } from './updateSupplierUserProfileDto';


export interface UpdateSupplierProfileDto { 
    supplier?: UpdateSupplierSupplierDto;
    userProfile?: UpdateSupplierUserProfileDto;
}